+++ internal/utils.py
@@ -1,4 +1,5 @@
 """stores utils that are shared between scripts"""
+import argparse
 import yaml
 
 with open("config.yml", "r") as ymlfile:
@@ -25,3 +26,27 @@ def get_covenant(args):
     else:
         covenants = []
     return covenants
+
+
+def get_simc_dir(talent, covenant, folder_name):
+    """get proper directory based on talent and covenant options"""
+    if covenant:
+        return "{0}/{1}/{2}/".format(folder_name, talent, covenant)
+    if talent:
+        return "{0}/{1}/".format(folder_name, talent)
+    return "{0}/".format(folder_name)
+
+
+def generate_parser(description):
+    """creates the shared argparser for sim.pu and profiles.py"""
+    parser = argparse.ArgumentParser(description=description)
+    parser.add_argument('dir', help='Directory to generate profiles for.')
+    parser.add_argument(
+        '--dungeons', help='Run a dungeonsimming batch of sims.', action='store_true')
+    parser.add_argument(
+        '--talents', help='indicate talent build for output.', choices=config["builds"].keys())
+    parser.add_argument(
+        '--covenant', help='indicate covenant build for output.', choices=config["covenants"])
+    parser.add_argument(
+        '--ptr', help='indicate if the sim should use ptr data.', action='store_true')
+    return parser
